product_review/requirements.txt
django==4.0
djangorestframework==3.13.1
django-cors-headers==3.10.1
djoser==2.1.0
pillow==8.4.0
